Hoist common OSProcess fields + add Solaris/AIX /proc base
Two cross-platform OSProcess dedups:

- Hoist the 13 uniform attribute fields (name, path, state, parentProcessID,
  threadCount, priority, virtualSize, kernelTime, userTime, startTime, upTime,
  bytesRead, bytesWritten) and their trivial getters into AbstractOSProcess.
  Declared protected (with a VisibilityModifier suppression) so each platform's
  updateAttributes() keeps assigning them directly. Removed from all six per-OS
  bases (Linux, Windows, macOS, BSD, Solaris, AIX). macOS keeps its historical
  empty-string name default via its constructor.

- Add AbstractProcOSProcess (oshi.software.common.os.unix) for the /proc-struct
  readers Solaris and AIX, sharing getCurrentWorkingDirectory, fd-count
  getOpenFiles, pflags getBitness, the /proc/<pid>/lwp thread skeleton (via a
  createThread hook; AIX now implements it concretely), the command-line backup
  fallback, and the user/group/resident-memory accessors.

Fixes a pre-existing AIX bug: getCurrentWorkingDirectory built "/proc" + pid
(missing the slash) instead of "/proc/" + pid; the shared base uses the correct
path. Left per-OS: updateAttributes, getAffinityMask, thread factories, args/env
plumbing, and the (genuinely divergent) state-char mappings.

Net -452 LOC. Validated locally on macOS via NativeComparisonTest (39/39,
JNA==FFM); Linux/Windows via PR CI, BSD/Solaris via unix.yaml, AIX via cfarm.
