+++ config/checkstyle-suppressions.xml
@@ -67,6 +67,12 @@
          FFM stat computations read identically. -->
     <suppress checks="MemberName" files="MacInternetProtocolStats\.java"/>
 
+    <!-- AbstractOSProcess holds the common OSProcess attribute fields as protected so each platform's
+         updateAttributes() can assign them directly (the alternative, private fields + protected setters,
+         would churn every native refresh method). -->
+    <suppress checks="VisibilityModifier" files="AbstractOSProcess\.java"/>
+    <suppress checks="VisibilityModifier" files="AbstractProcOSProcess\.java"/>
+
     <!-- BsdOSProcess/BsdOSThread/MacOSProcess are shared OSProcess/OSThread bases; their protected fields are
          populated by the per-backend native updateAttributes() in the subclasses (and the thread id / version are
          set by the per-platform constructors), so they can't be private with accessors. -->
