+++ oshi-common/src/main/java/oshi/software/common/os/unix/aix/AixOSProcess.java
@@ -5,76 +5,37 @@
 package oshi.software.common.os.unix.aix;
 
 import static oshi.software.os.OSProcess.State.INVALID;
-import static oshi.software.os.OSThread.ThreadFiltering.VALID_THREAD;
 import static oshi.util.Memoizer.defaultExpiration;
 import static oshi.util.Memoizer.memoize;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.unix.aix.AixLwpsInfo;
 import oshi.driver.common.unix.aix.AixPsInfo;
 import oshi.driver.common.unix.aix.PsInfo;
-import oshi.software.common.AbstractOSProcess;
+import oshi.software.common.os.unix.AbstractProcOSProcess;
 import oshi.software.os.OSThread;
 import oshi.util.Constants;
-import oshi.util.ExecutingCommand;
 import oshi.util.ParseUtil;
 import oshi.util.UserGroupInfo;
 import oshi.util.tuples.Pair;
 import oshi.util.tuples.Quartet;
 
 /**
- * Abstract base for AIX OSProcess. The /proc parsing, command-line/env memoization, thread enumeration, and field
- * assignment are shared; concrete subclasses (JNA/FFM) provide the perfstat lookup, the {@code rlimit} read, the
- * affinity supplier, and the actual {@code queryArgsEnv} address-space read.
+ * Abstract base for AIX OSProcess. The {@code /proc} parsing, command-line/env memoization, thread enumeration, and
+ * field assignment are shared (via {@link AbstractProcOSProcess}); concrete subclasses (JNA/FFM) provide the perfstat
+ * lookup, the {@code rlimit} read, the affinity supplier, and the actual {@code queryArgsEnv} address-space read.
  */
 @ThreadSafe
-public abstract class AixOSProcess extends AbstractOSProcess {
+public abstract class AixOSProcess extends AbstractProcOSProcess {
 
-    private static final Logger LOG = LoggerFactory.getLogger(AixOSProcess.class);
-
-    private final Supplier<Integer> bitnessSupplier = memoize(this::queryBitness);
     private final Supplier<AixPsInfo> psinfo = memoize(this::queryPsInfoMemo, defaultExpiration());
-    private final Supplier<String> commandLine = memoize(this::queryCommandLine);
-    private final Supplier<Pair<List<String>, Map<String, String>>> cmdEnv = memoize(this::queryCommandlineEnvironment);
-
-    private String name;
-    private String path = "";
-    private String commandLineBackup;
-    private String user;
-    private String userID;
-    private String group;
-    private String groupID;
-    private State state = State.INVALID;
-    private int parentProcessID;
-    private int threadCount;
-    private int priority;
-    private long virtualSize;
-    private long residentSetSize;
-    private long privateResidentMemory;
-    private long kernelTime;
-    private long userTime;
-    private long startTime;
-    private long upTime;
-    private long bytesRead;
-    private long bytesWritten;
 
     protected AixOSProcess(int pid) {
         super(pid);
@@ -85,76 +46,8 @@ public abstract class AixOSProcess extends AbstractOSProcess {
     }
 
     @Override
-    public String getName() {
-        return this.name;
-    }
-
-    @Override
-    public String getPath() {
-        return this.path;
-    }
-
-    @Override
-    public String getCommandLine() {
-        return this.commandLine.get();
-    }
-
-    private String queryCommandLine() {
-        String cl = String.join(" ", getArguments());
-        return cl.isEmpty() ? this.commandLineBackup : cl;
-    }
-
-    @Override
-    public List<String> getArguments() {
-        return cmdEnv.get().getA();
-    }
-
-    @Override
-    public Map<String, String> getEnvironmentVariables() {
-        return cmdEnv.get().getB();
-    }
-
-    private Pair<List<String>, Map<String, String>> queryCommandlineEnvironment() {
+    protected Pair<List<String>, Map<String, String>> queryCommandlineEnvironment() {
         return queryArgsEnv(getProcessID(), psinfo.get());
-    }
-
-    @Override
-    public String getCurrentWorkingDirectory() {
-        try {
-            String cwdLink = "/proc" + getProcessID() + "/cwd";
-            String cwd = new File(cwdLink).getCanonicalPath();
-            if (!cwd.equals(cwdLink)) {
-                return cwd;
-            }
-        } catch (IOException e) {
-            LOG.trace("Couldn't find cwd for pid {}: {}", getProcessID(), e.getMessage());
-        }
-        return "";
-    }
-
-    @Override
-    public String getUser() {
-        return this.user;
-    }
-
-    @Override
-    public String getUserID() {
-        return this.userID;
-    }
-
-    @Override
-    public String getGroup() {
-        return this.group;
-    }
-
-    @Override
-    public String getGroupID() {
-        return this.groupID;
-    }
-
-    @Override
-    public State getState() {
-        return this.state;
     }
 
     /**
@@ -170,91 +63,8 @@ public abstract class AixOSProcess extends AbstractOSProcess {
     }
 
     @Override
-    public int getParentProcessID() {
-        return this.parentProcessID;
-    }
-
-    @Override
-    public int getThreadCount() {
-        return this.threadCount;
-    }
-
-    @Override
-    public int getPriority() {
-        return this.priority;
-    }
-
-    @Override
-    public long getVirtualSize() {
-        return this.virtualSize;
-    }
-
-    @Override
-    public long getResidentMemory() {
-        return this.residentSetSize;
-    }
-
-    @Override
-    public long getPrivateResidentMemory() {
-        return this.privateResidentMemory;
-    }
-
-    @Override
-    public long getKernelTime() {
-        return this.kernelTime;
-    }
-
-    @Override
-    public long getUserTime() {
-        return this.userTime;
-    }
-
-    @Override
-    public long getUpTime() {
-        return this.upTime;
-    }
-
-    @Override
-    public long getStartTime() {
-        return this.startTime;
-    }
-
-    @Override
-    public long getBytesRead() {
-        return this.bytesRead;
-    }
-
-    @Override
-    public long getBytesWritten() {
-        return this.bytesWritten;
-    }
-
-    @Override
-    public long getOpenFiles() {
-        try (Stream<Path> fd = Files.list(Paths.get("/proc/" + getProcessID() + "/fd"))) {
-            return fd.count();
-        } catch (IOException e) {
-            return 0L;
-        }
-    }
-
-    @Override
-    public int getBitness() {
-        return this.bitnessSupplier.get();
-    }
-
-    private int queryBitness() {
-        List<String> pflags = ExecutingCommand.runNative("pflags " + getProcessID());
-        for (String line : pflags) {
-            if (line.contains("data model")) {
-                if (line.contains("LP32")) {
-                    return 32;
-                } else if (line.contains("LP64")) {
-                    return 64;
-                }
-            }
-        }
-        return 0;
+    protected OSThread createThread(int lwpid) {
+        return new AixOSThread(getProcessID(), lwpid);
     }
 
     @Override
@@ -274,19 +84,6 @@ public abstract class AixOSProcess extends AbstractOSProcess {
         }
         mask &= getCpuAffinityMask();
         return mask;
-    }
-
-    @Override
-    public List<OSThread> getThreadDetails() {
-        File directory = new File(String.format(Locale.ROOT, "/proc/%d/lwp", getProcessID()));
-        File[] numericFiles = directory.listFiles(file -> Constants.DIGITS.matcher(file.getName()).matches());
-        if (numericFiles == null) {
-            return Collections.emptyList();
-        }
-        return Arrays.stream(numericFiles).parallel()
-                .map(lwpidFile -> (OSThread) new AixOSThread(getProcessID(),
-                        ParseUtil.parseIntOrDefault(lwpidFile.getName(), 0)))
-                .filter(VALID_THREAD).collect(Collectors.toList());
     }
 
     /**
