+++ oshi-common/src/main/java/oshi/software/common/os/unix/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2026 The OSHI Project Contributors
+ * SPDX-License-Identifier: MIT
+ */
+/**
+ * Provides shared abstract base classes for the {@code /proc}-filesystem Unix operating systems
+ */
+package oshi.software.common.os.unix;
